Added access to top-level summaries and icons
The Daily, Hourly, and Minutely data blocks each have a top-level summary and icon property. These give an overview of the entire time period and are not tied to a specific day / hour / minute.

These summaries are part of the FIODataBlock class, but were not accessible from the FIODaily, FIOHourly, or FIOMinutely classes.

Added getter methods to these three classes to provide access.