Added methods to get the FIODataBlock
Added getter methods to get the FIODataBlock object for Daily, Hourly, and Minutely. This is useful if you need to write a single loop to check for particular data in all three. For example, you want to know when it will rain next, so you want to loop through each forecast (minutely, hourly, and daily) checking for precipitation and stopping when it is found. This can all be done using the same loop if you look at the FIODataBlock object.

The specific properties available in the block do vary depending on the type of forecast, so be careful about which properties you attempt to check.

Also fixed comments on my earlier commit for the top-level summaries.